Refactor: 가게 등록  input 컴포넌트 분리
