+++ src/components/myInfoComponents/assignMyShopInfo/shopAddress/index.tsx
@@ -1,0 +1,40 @@
+import DropDown from '@/components/commonComponents/dropdown';
+import Input from '@/components/commonComponents/input';
+import LOCATION from '@/constant/location';
+import useStoreShopInfo from '@/stores/storeShopInfo';
+import { ChangeEvent } from 'react';
+
+const ShopAddress = () => {
+  const { setShopData } = useStoreShopInfo();
+
+  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    setShopData({ [name]: value });
+  };
+
+  const handleDropDownChange = (name: string, value: string) => {
+    setShopData({ [name]: value });
+  };
+
+  return (
+    <div className="w-[100%]  flex flex-col md:flex-row gap-[20px]">
+      <div className="flex flex-col items-start gap-[8px] w-[100%]">
+        <p>주소</p>
+        <DropDown
+          menuItems={LOCATION}
+          className="w-[100%] bg-white h-[58px] border rounded-[6px] border-gray30 py-[16px] px-[20px]"
+          onSelect={(value) => handleDropDownChange('address1', value)}
+        />
+      </div>
+      <Input
+        className="w-[100%] "
+        variant="normal"
+        name="address2"
+        label="상세주소"
+        onChange={handleInputChange}
+      />
+    </div>
+  );
+};
+
+export default ShopAddress;
